front: move old allowances to stdcm, to be rewritten very shortly
