+++ front/src/applications/stdcm/components/OldAllowances/StandardAllowanceDefault.tsx
