+++ front/src/applications/stdcm/components/OldAllowances/allowancesConsts.tsx
